Core: Fixes for various unsafe implicit type conversion warnings in the Core module (WIP; some files still suppress the warning) - Ptrdiff -> int32 - Float/int confusion and double/float - size_t stuff; various changes to the algorithms to use a deduced IndexType template argument and/or decltype to use the appropriate size for indicies and counts - Fixed GetNum(FString) incorrectly returning SIZE_T instead of int32, and GetNum(container) now returns whatever container.Num() does (so usually int32) #jira UE-86949 #rb marc.audy, steve.robb
#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: michael.noland
#ROBOMERGE-SOURCE: CL 11050799 via CL 11050828 via CL 11050837 via CL 11051763
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v640-11091645)

[CL 11127544 by michael noland in Dev-Networking branch]

+++ Engine/Source/Runtime/Core/Private/Misc/AutomationTest.cpp
@@ -164,7 +164,7 @@ bool FAutomationTestFramework::RunSmokeTests()
 			TMap<FString, FAutomationTestExecutionInfo> OutExecutionInfoMap;
 
 			// Run each valid test
-			FScopedSlowTask SlowTask(TestInfo.Num());
+			FScopedSlowTask SlowTask((float)TestInfo.Num());
 
 			// We disable capturing the stack when running smoke tests, it adds too much overhead to do it at startup.
 			FAutomationTestFramework::Get().SetCaptureStack(false);
