+++ Engine/Source/Runtime/Core/Private/ProfilingDebugging/CpuProfilerTrace.cpp
@@ -139,7 +139,7 @@ void FCpuProfilerTraceInternal::EndCapture(FThreadBuffer* InThreadBuffer)
 	uint8* BufferPtr = ThreadBuffer->Buffer + ThreadBuffer->BufferSize; \
 	FTraceUtils::Encode7bit((CycleDiff << 1) | 1ull, BufferPtr); \
 	FTraceUtils::Encode7bit(SpecId, BufferPtr); \
-	ThreadBuffer->BufferSize = BufferPtr - ThreadBuffer->Buffer; \
+	ThreadBuffer->BufferSize = (uint16)(BufferPtr - ThreadBuffer->Buffer); \
 	if (ThreadBuffer->BufferSize >= FCpuProfilerTraceInternal::FullBufferThreshold) \
 	{ \
 		FCpuProfilerTraceInternal::FlushThreadBuffer(ThreadBuffer); \
@@ -208,7 +208,7 @@ void FCpuProfilerTrace::OutputEndEvent()
 	ThreadBuffer->LastCycle = Cycle;
 	uint8* BufferPtr = ThreadBuffer->Buffer + ThreadBuffer->BufferSize;
 	FTraceUtils::Encode7bit(CycleDiff << 1, BufferPtr);
-	ThreadBuffer->BufferSize = BufferPtr - ThreadBuffer->Buffer;
+	ThreadBuffer->BufferSize = (uint16)(BufferPtr - ThreadBuffer->Buffer);
 	if ((FCpuProfilerTraceInternal::ThreadDepth == 0) | (ThreadBuffer->BufferSize >= FCpuProfilerTraceInternal::FullBufferThreshold))
 	{
 		FCpuProfilerTraceInternal::FlushThreadBuffer(ThreadBuffer);
